requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23

# Database drivers
psycopg2-binary==2.9.9

# Environment configuration
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Azure SDK (optional, for Azure-specific features)
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Utilities
Werkzeug==3.0.1
